+++ src/pages/test-components/index.tsx
@@ -20,8 +20,8 @@ import SheetSection from "@/components/sheet/sheet-section";
 import Typography from "@/components/typography";
 import { CommonSize } from "@/constants/common/common-size";
 import { ProfileRandomColor } from "@/constants/profile-random-color";
-import ImageInput from "@/features/edit-task/components/image-input";
-import TagInput from "@/features/edit-task/components/tag-input";
+import ImageInput from "@/features/card/components/image-input";
+import TagInput from "@/features/card/components/tag-input";
 import { useAlert } from "@/hooks/use-alert";
 import { useDialog } from "@/hooks/use-dialog";
 import { useModal } from "@/hooks/use-modal";
